Added logic for authorized update and delete

+++ src/routes/postRoutes.js
@@ -36,35 +36,53 @@ router.get("/",async (req,res)=>{
 });
 
 router.put("/:id",async(req,res)=>{
-    try{
-        await post.findByIdAndUpdate(req.params.id,req.body);
-        res.status(200).json({
-            status:"Success",
-            message:"Post updated"
-        })
-
-    }catch(e){
-        res.status(400).json({
+    const data = await post.findById(req.params.id);
+    if(data.user.toString() != req.user.toString()){
+        res.status(404).json({
             status:"failed",
-            message:e.message
+            message:"Not Authorized to change"
         })
+    }else{
+        try{
+            await post.findByIdAndUpdate(req.params.id,req.body);
+            res.status(200).json({
+                status:"Success",
+                message:"Post updated"
+            })
+    
+        }catch(e){
+            res.status(400).json({
+                status:"failed",
+                message:e.message
+            })
+        }
     }
+   
     
 });
 
 router.delete("/:id",async(req,res)=>{
-    try{
-        await post.findByIdAndDelete(req.params.id);
-        res.status(200).json({
-            status:"Success",
-            message:"Successfully Deleted"
-        })
-    }catch(e){
-        res.status(400).json({
+    const data = await post.findById(req.params.id);
+    if(data.user.toString() != req.user.toString()){
+        res.status(404).json({
             status:"failed",
-            message:e.message
+            message:"Not Authorized to delete"
         })
+    }else{
+        try{
+            await post.findByIdAndDelete(req.params.id);
+            res.status(200).json({
+                status:"Success",
+                message:"Successfully Deleted"
+            })
+        }catch(e){
+            res.status(400).json({
+                status:"failed",
+                message:e.message
+            })
+        }
     }
+   
 })
 
 module.exports = router;
